pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "longcat-design"
version = "0.1.0"
description = "Layered poster generation pipeline: agent + NBP + editable PSD/SVG output + training-ready trajectories."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.40.0",
  "google-genai>=1.0.0",
  "Pillow>=10.0.0",
  "psd-tools>=1.11.0",
  "svgwrite>=1.4.3",
  "pydantic>=2.5.0",
  "python-dotenv>=1.0.0",
  "fonttools>=4.50.0",
  "brotli>=1.1.0",
]

[project.scripts]
longcat-design = "longcat_design.cli:main"

[tool.setuptools.packages.find]
include = ["longcat_design*"]
